These changes aren't needed yet

+++ src/main/java/moze_intel/projecte/gameObjs/gui/GUIRelayMK1.java
@@ -53,7 +53,7 @@ public class GUIRelayMK1 extends GuiContainer
 		this.drawTexturedModalRect(x, y, 0, 0, xSize, ySize);
 		
 		//Emc bar progress. Max is 102.
-		int progress = (int) ((double) container.emc / tile.getMaximumEmc() * 102);
+		int progress = (int) (container.emc / tile.getMaximumEmc() * 102);
 		this.drawTexturedModalRect(x + 64, y + 6, 30, 177, progress, 10);
 		
 		//Klein start bar progress. Max is 30.
